Move the base Linker class
